core: fe refactor
