+++ .env.local
@@ -36,13 +36,16 @@ RUST_HOME_LIMIT_MS=50
 SVELTE_PORT=5173
 SVELTE_NAME=svelte
 SVELTE_URL=https://${CADDY_DOMAIN}
+
+# Svelte Swap
 PUBLIC_SVELTE_SWAP_ROOT=/swap
-PUBLIC_SVELTE_ATTEND_ROOT=/attend
 PUBLIC_PAGE_SIZE=10
 PUBLIC_MEILI_KEY=2a32c3edd5376c1c3bf95b972f1717abe669ca91a6b85a8fd7a6e3a1012dbcc8
 PUBLIC_MEILI_PATH=/search
 PUBLIC_MEILI_URL=https://${CADDY_DOMAIN}${PUBLIC_MEILI_PATH}
 PUBLIC_SWAP_BACKEND_PATH=${PUBLIC_SVELTE_SWAP_ROOT}/api
+
+# Svelte Other
 PUBLIC_HOME_BACKEND_PATH=/home/api
 
 # Meilisearch
