+++ frontend/Dockerfile
@@ -11,11 +11,9 @@ ARG PUBLIC_MEILI_URL
 ARG PUBLIC_PAGE_SIZE
 ARG PUBLIC_MIN_PASSWORD_LENGTH
 ARG PUBLIC_SVELTE_SWAP_ROOT
-ARG PUBLIC_SVELTE_ATTEND_ROOT
 ARG PORT
 
 ENV NODE_ENV=production
-ENV PUBLIC_SVELTE_ATTEND_ROOT=$PUBLIC_SVELTE_ATTEND_ROOT
 ENV PUBLIC_MIN_PASSWORD_LENGTH=$PUBLIC_MIN_PASSWORD_LENGTH
 ENV PUBLIC_SWAP_BACKEND_PATH=$PUBLIC_SWAP_BACKEND_PATH
 ENV PUBLIC_HOME_BACKEND_PATH=$PUBLIC_HOME_BACKEND_PATH
