+++ frontend/src/lib/swap/components/auth/subcomponents/Email.svelte
