+++ frontend/src/lib/swap/components/auth/subcomponents/Password.svelte
