+++ frontend/src/lib/swap/components/faq/Header.svelte
@@ -1,0 +1,6 @@
+<script lang="ts"></script>
+
+<div class="text-center mb-12">
+	<h1 class="text-4xl font-bold text-gray-900 mb-4">? Frequently Asked Questions</h1>
+	<p class="text-xl text-gray-600">Find quick answers to common questions about BoilerSwap</p>
+</div>
