+++ frontend/src/lib/swap/components/how-it-works/Guide.svelte
@@ -1,0 +1,40 @@
+<script lang="ts"></script>
+
+<div class="space-y-12">
+	<section>
+		<h2 class="text-2xl font-bold text-gray-900 mb-6">For Item Givers</h2>
+		<div class="grid grid-cols-1 gap-6">
+			<div class="bg-white rounded-lg shadow-sm border p-6">
+				<h3 class="text-xl font-semibold mb-4 flex items-center">
+					<span class="text-2xl mr-3">📤</span>
+					Posting Items
+				</h3>
+				<ul class="space-y-2 text-gray-600">
+					<li>• Choose from predefined item categories</li>
+					<li>• Rate the condition honestly</li>
+					<li>• Add an optional description</li>
+					<li>• Set pickup location</li>
+					<li>• Items auto-expires after a week</li>
+				</ul>
+			</div>
+		</div>
+	</section>
+
+	<section>
+		<h2 class="text-2xl font-bold text-gray-900 mb-6">For Item Seekers</h2>
+		<div class="grid grid-cols-1 gap-6">
+			<div class="bg-white rounded-lg shadow-sm border p-6">
+				<h3 class="text-xl font-semibold mb-4 flex items-center">
+					<span class="text-2xl mr-3">🔍</span>
+					Finding Items
+				</h3>
+				<ul class="space-y-2 text-gray-600">
+					<li>• Browse by category, location, or condition</li>
+					<li>• Use search to find specific items</li>
+					<li>• See item condition</li>
+					<li>• Check pickup locations</li>
+				</ul>
+			</div>
+		</div>
+	</section>
+</div>
