+++ frontend/src/lib/swap/components/how-it-works/Header.svelte
@@ -1,0 +1,6 @@
+<script lang="ts"></script>
+
+<div class="text-center mb-16">
+	<h1 class="text-4xl font-bold text-gray-900 mb-4">How BoilerSwap Works</h1>
+	<p class="text-xl text-gray-600">Two simple steps to reduce waste and help fellow Boilermakers</p>
+</div>
