+++ frontend/src/lib/swap/components/how-it-works/Instructions.svelte
@@ -1,0 +1,21 @@
+<script lang="ts"></script>
+
+<div class="grid grid-cols-1 md:grid-cols-2 gap-8 mb-16">
+	<div class="text-center bg-white p-6 rounded-lg shadow-sm">
+		<div class="w-16 h-16 bg-yellow-100 rounded-full flex items-center justify-center mx-auto mb-4">
+			<span class="text-2xl">📤</span>
+		</div>
+		<h3 class="text-xl font-semibold mb-2">1. Post Your Item</h3>
+		<p class="text-gray-600">
+			Pick an emoji, add a title, and set your location. It takes less than 2 minutes!
+		</p>
+	</div>
+
+	<div class="text-center bg-white p-6 rounded-lg shadow-sm">
+		<div class="w-16 h-16 bg-yellow-100 rounded-full flex items-center justify-center mx-auto mb-4">
+			<span class="text-2xl">🔍</span>
+		</div>
+		<h3 class="text-xl font-semibold mb-2">2. Students Browse</h3>
+		<p class="text-gray-600">Other students browse, search, and find items they need.</p>
+	</div>
+</div>
