+++ frontend/src/lib/swap/components/posting/PostingHeader.svelte
@@ -1,0 +1,6 @@
+<script lang="ts"></script>
+
+<div class="text-center mb-8">
+	<h1 class="text-3xl font-bold text-gray-900 mb-4">Post a New Item</h1>
+	<p class="text-gray-600">Help a fellow Boilermaker and keep usable items out of the trash!</p>
+</div>
