--- frontend/src/routes/attend/+page.svelte
@@ -1,13 +1,0 @@
-<script>
-</script>
-
-<main class="flex items-center justify-center h-screen bg-gray-100">
-	<div class="flex space-x-4">
-		<button class="px-6 py-3 bg-blue-600 text-white rounded-lg hover:bg-blue-700">
-			Button 1
-		</button>
-		<button class="px-6 py-3 bg-green-600 text-white rounded-lg hover:bg-green-700">
-			Button 2
-		</button>
-	</div>
-</main>
