+++ frontend/src/routes/swap/+layout.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import NavBar from '$lib/swap/components/layout/NavBar.svelte'
 	import Footer from '$lib/swap/components/layout/Footer.svelte'
-	import { PUBLIC_SVELTE_SWAP_ROOT } from '$env/static/public'
+	import PostButton from '$lib/swap/components/layout/PostButton.svelte'
 	import { appState } from '$lib/swap/AppState.svelte'
 	import { Status } from '$lib/swap/models'
 	import { onDestroy, onMount } from 'svelte'
@@ -15,7 +15,7 @@
 	onMount(() => {
 		refreshToken()
 
-		interval = setInterval(refreshToken, 270000)
+		interval = setInterval(refreshToken, 270_000)
 	})
 
 	onDestroy(() => {
@@ -47,10 +47,5 @@
 {/if}
 
 {#if !page.url.pathname.includes('/auth') && !page.url.pathname.includes('/post')}
-	<a
-		href={PUBLIC_SVELTE_SWAP_ROOT + '/post'}
-		class="fixed bottom-6 right-6 w-14 h-14 rounded-full bg-gradient-to-r from-yellow-400 to-amber-500 hover:from-yellow-500 hover:to-amber-600 text-white shadow-lg transition-all hover:scale-110 flex items-center justify-center"
-	>
-		<span class="text-2xl">+</span>
-	</a>
+	<PostButton />
 {/if}
