+++ frontend/src/routes/swap/auth/+page.svelte
@@ -1,16 +1,8 @@
 <script lang="ts">
-	import { TabOptionsIterable, type TabOptions } from '$lib/swap/models'
 	import AuthHeader from '$lib/swap/components/auth/AuthHeader.svelte'
-	import TabButton from '$lib/swap/components/auth/TabButton.svelte'
-	import AuthFlow from '$lib/swap/components/auth/AuthFlow.svelte'
 	import { onMount } from 'svelte'
 	import { signout } from '$lib/swap/helpers/auth'
-
-	let activeTab: TabOptions = $state('Login')
-
-	function showTab(tab: TabOptions) {
-		activeTab = tab
-	}
+	import AuthInput from '$lib/swap/components/auth/AuthInput.svelte'
 
 	onMount(() => {
 		signout()
@@ -19,22 +11,8 @@
 
 <svelte:head>
 	<title>Login - BoilerSwap</title>
+	<meta name="description" content="Sign in, create an account, or change your password." />
 </svelte:head>
 
 <AuthHeader />
-
-<div class="container mx-auto px-6 py-16 max-w-md">
-	<div class="text-center mb-8">
-		<h1 class="text-3xl font-bold text-gray-900 mb-4">Welcome to BoilerSwap</h1>
-		<p class="text-gray-600">Join the Purdue community of sustainable sharing</p>
-	</div>
-
-	<div class="bg-white rounded-lg shadow-sm border p-6">
-		<div class="flex border-b mb-6">
-			{#each TabOptionsIterable as tab}
-				<TabButton bind:activeTabValue={activeTab} tabName={tab} {showTab} />
-			{/each}
-		</div>
-		<AuthFlow bind:activeTabValue={activeTab} {showTab} />
-	</div>
-</div>
+<AuthInput />
