+++ frontend/src/routes/swap/auth/verify/update/+page.svelte
@@ -1,17 +1,11 @@
 <script lang="ts">
 	import { goto } from '$app/navigation'
-	import {
-		PUBLIC_TEMP_SESSION_DURATION_SECS,
-		PUBLIC_MAX_CHARS,
-		PUBLIC_MIN_PASSWORD_LENGTH,
-		PUBLIC_SVELTE_SWAP_ROOT
-	} from '$env/static/public'
+	import { PUBLIC_TEMP_SESSION_DURATION_SECS, PUBLIC_SVELTE_SWAP_ROOT } from '$env/static/public'
 	import { appState } from '$lib/swap/AppState.svelte'
+	import UpdatePassword from '$lib/swap/components/auth/verify/UpdatePassword.svelte'
 	import { Status } from '$lib/swap/models'
 	import { onDestroy, onMount } from 'svelte'
-	import { update } from '$lib/swap/helpers/auth'
 
-	let new_password: string = $state('')
 	let timer: number | null = null
 
 	onMount(() => {
@@ -41,31 +35,6 @@
 
 <div class="container mx-auto px-6 py-16 max-w-md">
 	<div class="bg-white rounded-lg shadow-sm border p-6">
-		<form onsubmit={() => update(new_password)} class="space-y-4">
-			<p class="text-gray-600 text-sm">Enter your new password.</p>
-			<div>
-				<label class="block text-sm font-medium mb-2">
-					New Password
-					<input
-						type="text"
-						bind:value={new_password}
-						placeholder="abcdefghijklmnop"
-						class="w-full px-4 py-2 border rounded-lg"
-						maxlength={Number(PUBLIC_MAX_CHARS)}
-						minlength={Number(PUBLIC_MIN_PASSWORD_LENGTH)}
-						required
-					/>
-				</label>
-				<p class="text-xs text-gray-500 mt-1">Must be {PUBLIC_MIN_PASSWORD_LENGTH}+ characters</p>
-			</div>
-			<button
-				type="submit"
-				class="w-full {appState.getLimited()
-					? 'bg-gray-300 cursor-not-allowed'
-					: 'bg-yellow-400 hover:bg-yellow-500'} text-gray-800 py-2 rounded-lg transition-colors"
-			>
-				Submit
-			</button>
-		</form>
+		<UpdatePassword />
 	</div>
 </div>
