+++ frontend/src/routes/swap/faq/+page.svelte
@@ -1,50 +1,14 @@
 <script lang="ts">
+	import Header from '$lib/swap/components/faq/Header.svelte'
+	import Questions from '$lib/swap/components/faq/Questions.svelte'
 </script>
 
 <svelte:head>
 	<title>FAQ - BoilerSwap</title>
+	<meta name="description" content="Frequently asked questions about BoilerSwap." />
 </svelte:head>
 
 <div class="container mx-auto px-6 py-16 max-w-4xl">
-	<div class="text-center mb-12">
-		<h1 class="text-4xl font-bold text-gray-900 mb-4">? Frequently Asked Questions</h1>
-		<p class="text-xl text-gray-600">Find quick answers to common questions about BoilerSwap</p>
-	</div>
-
-	<div class="bg-white rounded-lg shadow-sm border p-8">
-		<h2 class="text-2xl font-bold mb-6">🚀 Getting Started</h2>
-		<div class="space-y-6">
-			<div class="border-b pb-4">
-				<h3 class="text-lg font-semibold mb-2">How do I create an account?</h3>
-				<p class="text-gray-600">
-					You need a valid @purdue.edu email address to sign up. Click "Sign In" in the header, then
-					"Sign Up" and follow the verification process.
-				</p>
-			</div>
-
-			<div class="border-b pb-4">
-				<h3 class="text-lg font-semibold mb-2">Is BoilerSwap really free?</h3>
-				<p class="text-gray-600">
-					Yes! BoilerSwap is completely free to use. All items shared on the platform are given away
-					for free by fellow students.
-				</p>
-			</div>
-
-			<div class="border-b pb-4">
-				<h3 class="text-lg font-semibold mb-2">Who can use BoilerSwap?</h3>
-				<p class="text-gray-600">
-					BoilerSwap is exclusively for Purdue University students, faculty, and staff with valid
-					@purdue.edu email addresses.
-				</p>
-			</div>
-
-			<div class="pb-4">
-				<h3 class="text-lg font-semibold mb-2">How do I verify my Purdue email?</h3>
-				<p class="text-gray-600">
-					After signing up, we'll send a verification code to your @purdue.edu email. Enter in the
-					code to activate your account.
-				</p>
-			</div>
-		</div>
-	</div>
+	<Header />
+	<Questions />
 </div>
