+++ frontend/src/routes/swap/post/+page.svelte
@@ -1,17 +1,16 @@
 <script lang="ts">
 	import PostingForm from '$lib/swap/components/posting/PostingForm.svelte'
+	import PostingHeader from '$lib/swap/components/posting/PostingHeader.svelte'
 	import PostingTips from '$lib/swap/components/posting/PostingTips.svelte'
 </script>
 
 <svelte:head>
 	<title>Post Item - BoilerSwap</title>
+	<meta name="description" content="Posting new item." />
 </svelte:head>
 
 <div class="container mx-auto px-6 py-8 max-w-2xl">
-	<div class="text-center mb-8">
-		<h1 class="text-3xl font-bold text-gray-900 mb-4">Post a New Item</h1>
-		<p class="text-gray-600">Help a fellow Boilermaker and keep usable items out of the trash!</p>
-	</div>
+	<PostingHeader />
 	<PostingForm />
 	<PostingTips />
 </div>
